Cleaner Code & Added Functionality
Cleaned up duplicate code in the css files, and made contact.css Added a contact strip in the index page. Made the projects panels clickable. Added an icons images folder. Changed the about on the index page to be shorter and have a read more button. Changed the header height to look more neat.